Duplicate siunitx.dtx as siunitx-v2.dtx
This is the first step in the v3 rewrite. The plan here is to recode
the mechanisms a little at a time into various new files. Thus from
here 'onward' the main branch will most likely not be useable for some
time (until the various pieces of work are done and everything is
ready to be reassembled).