Remove 'early' loading of expl3/xparse
The driver is using `l3doc` so this is redundant anyway (`xparse` should
be loaded 'normally' if/when needed). Loading `expl3` can happily be
covered in the 'proper' code sections as everything is built up, and
is only currently appropriate in the main file anyway (assuming DocStrip
is used to make a single `.sty`).