Define a test for literal versus symbolic units
In contrast to the v2 test, the plan here is that the symbolic units
will be expandable, at least to one 'level'. That should allow a quick
switch-based test rather than the current mapping approach.

To support this test, a switch is created which will be used by the
symbolic unit definition.

A local scratch token list is also defined: note that this is
submodule-specific, and so the code here will not directly depend on
similar scratch allocation in the other parts of the bundle.