Add routines to create symbolic prefixes/powers/qualifiers/units
The general mechanism here is slightly simpler that that in v2: a
single function does all of the definitions, with each of the symbolic
types feeding into this. In contrast to v2, there is no treatment of
unit options at this stage and prefixes are only stored with 'reverse
data' if they are powers of 10. Options will be handled at a 'higher
level', while the prefix conversion code is likely to be simplified to
stick more closely to what seems reasonable to handle. Also notice that
this approach uses only a single internal function to store both the
literal and symbolic input: the saving of tokens is not what is
crucial here, but rather the fact that using an additional switch should
keep mappings down and so improve performance over all.