Define various 'special' symbolic unit commands
This creates the non-standard commands, all of which are 'internal' to
the parser and so need to be set up here (not outside of the parser
block).

At the moment, `\cancel` might not be quite right: using the same name
twice might need some checks to avoid an infinite loop. That can get
tested properly later on.