Add a routine to do formatting of literal units
As it stands, this approach differs from that in v2 in that it assumes
that any symbolic units present are `x`-type expandable. This is
required so that a formatted token list can be returned outside of
the group that the parser will be contained within.

Notice that this uses the LaTeX2e-dependent `\protected@edef` as
explained in the comments.

This requires one general option be defined: the product separator
between units. At present this is stored in a sub-tree of the
main `siunitx` keys tree, although the name is unchanged from v2. The
plan here is to separate out the keys, again so that in the future
the various parts can be converted into independent modules. Later a
key mapping will be required.