Main parsed unit formatter refactored/copied
This version covers most of the functionality of the v2 one, with a
few odds-and-ends yet to tidy up. There are at least a couple of changes
to behaviour, which will all have to be checked later on (when updating
the docs). Notice that most of the equivalent code is removed from the
v2 file, with the bits remaining clearly 'to do'.

At the moment, this code does not cover the need to deal with the font
applied to the output. As that is being heavily revised for v3, it
makes sense to first check in something that ignores this area, then
to update the code to address it.