Transfer base units and SI prefixes to units submodule
It seems most sensible to define these here: looking forward, I'd
expect such definitions to go with a 'stand alone' units module rather
than being tied to the 'higher level' processing of quantities.

Note that the question of how to deal with non-ASCII symbols is still
open. The commands here really should not be 'design level', so for
the moment I have gone for `\unitsymbol` as a generic name.