Add named units, etc.
This covers most of the BIPM-approved stuff, with a few common
powers too. This all needs documenting yet, and there are the small
number of 'physical' units to worry about.

Open questions here are the exact detail of handling symbols (still
not entirely clear), what to do about capitalisation (Celsius and mmHg),
whether 'mmHg' is really an abbreviation, and the percent issue (is it
a unit, does it belong in here at all?)