Use \mathit for natural/atomic units
Simply using \ensuremath won't help for example inside \mathrm. As the
natural units are by convention in italic even though other units are
not, the cleanest way to handle this is to force the situation. An
alternative is to \mbox everything, but that leads to mode shifting and
is awkward. Another approach would be to choose \mathit/\textit based
on math/text mode 'higher up', but this gets very messy with the
interface and is probably not needed. In any case, these units are
pretty clear-cut 'mathematical'.