Drop "space" option for "unit/qualifier-mode", tidy up "combine" code
This should now all work properly. Note that as it stands there is an
issue with line breaking within a qualifier. As everything ends up in
the argument to e.g. `\mathrm` it's not possible to break before and
after any text at spaces. It's not clear how this might be handled:
even if the different parts are wrapped in separate font commands the
problem would not disappear. May need to rethink this entire area:
there is also a potential issue with a phrase if fonts differ. Perhaps
best to drop that and know a font group can be forced. (That all links
to the question of exactly how much to wrap in one font command.)