For the present, stick to "_"/"^" for sub/superscripts
This requires a bit of catcode trickery but means that the output is
more 'normal' in LaTeX terms. The reasoning here is that with the
change to how fonts are handled we can be sure that math mode is active
at the 'outer' level. That should mean that we can safely use math mode
for more things. There might still be something to worry about when we
come to text mode numbers, but we'll cross that bridge when we come to
it.