Add "number/input/..." options
A couple of options from v2 are removed here. The idea of protecting
tokens I think is actually not needed if the expansion step is
improved (LaTeX2e so \protected@edef is required). The option for
symbols is also going: the idea here is that if people really want to do
non-standard input it should either not be parsed or be treated as a
'number' (at the risk of the user).