Output decimal marker correctly
There's an open question here: how to deal with "," as a decimal
separator. In v2, an extra set of braces are added as part of the
storage. Probably that should stay in teh v2 file with in v3 users
expected to set the key 'correctly' if that is what they want.