Use low-level box closing
The recent expl3 update to colour-safe boxes means we now get
a breakage here. It's not really right to match a raw box
opening with a higher-level closing anyway, so for the moment
do both at a low level. This should be addressed by adding an
interface to expl3 for boxes-to-width with :w parsing.