Activation "print/follow-shape"
At present this may be too limited (do we need to make the 'fixed'
outcome available as a key, for example), and may need to be set up to
store the result, etc. However, this will need to be examined once a
fuller set of control systems are set up.