Drop "unit/font-command"
Working on the print code, it becomes clear that this doesn't really
help in any way. If one is working on the basis that the formatter
produces 'standard' math mode LaTeX, then it has to be \mathrm. Making
that a variable is extremely awkward as it has to be tided up at the
print stage in a problematic way.

The same therefore applies to the print submodule: the font setting is
removed there also.