Swtich back to (largely) 'flat' key structure
The idea of subdividing is attractive from a programming POV but gets
very awkward as a user: this became clear in starting the user docs. It
probably does make sense to key keys *that have limited context* in a
subtree, for example those for tables. (The latter have not been moved
but may be: will depend on how it looks at the point of writing the
docs.)

As a result of these changes, a small number of keys have to be renamed.