Error units package is loaded
In v2 this one was allowed (if odd): for v3.0 that did not
change, but as both now define \unit this is an issue. Since
siunitx and units are both unit packages, having \unit from
units and e.g. \qty from siunitx is an issue: here it is
better to block and get users to update their source than
to try to 'live with' the issue (cf. physics package considerations).