Switch from f- to e-type expansion
Came up in latex3/latex3#1257: a good test is
the expansion code in siunitx. Time impact on
running the test suite for pdfTeX is <0.5 s, which
is acceptable.

As part of this commit, some \exp_args:N...
are replaced by variants: I will clean up the rest in
a second commit.